+++ js/data.js
@@ -1,27 +1,3 @@
-/* eslint-disable no-console */
-/* eslint-disable eqeqeq */
-function getRandomInt(min, max) {
-  if (min <= max && min >= 0) {
-    return Math.floor(min + Math.random() * (max + 1 - min));
-  }
-  return 'Неверный диапазон чисел';
-}
-
-getRandomInt(0, 100);
-// Проверка.
-// console.log(getRandomInt(0, 100));
-
-function stringLength (text, peak = 40) {
-  if (text.length <= peak) {
-    return(true);
-  }
-  return(false);
-}
-
-stringLength('Любая строка короче ста символов');
-// Проверка.
-// console.log (stringLength('Любая строка больше ста символов (Любая строка больше ста символов)'));
-
 const NAMES = ['Артём', 'Дмитрий', 'Иван', 'Сергей', 'Алексей', 'Иван',];
 const DESCRIPTION = ['Звёздная ночь', 'Последний день Помпеи', 'Грачи прилетели', 'Мокрый луг', 'Девятый вал'];
 const MESSAGE = ['Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',];
