+++ js/main.js
@@ -1,2 +1,3 @@
 
 import './data.js';
+import './util.js';
